Use slowEmit when the file doesn't exist yet
fastEmit does a getFIle itself, en when the file doesn't exist
it throws an error. It expects that the file has been added or updated
before running the fastEmit.

In the case where files are required using the require.context glob, the
situation above happens, so we need to use the slowEmit path.